Use codecs package for file writing.
Use codecs to write files - so encoding happens for all strings written.
Mark strings with u for correct writing.

Function __cell_range to check for maximum number of rows - so we do not
try to access cells 'after' the end of the sheet.

Easier check for existence of read data - remove len() construct.

Add some comments.